backend/requirements.txt
fastapi
uvicorn
python-jose
pydantic
python-multipart